load by default editor.js, move new toolbar to toolbar.js
