+++ frontend/src/toolbar/index.js
@@ -1,4 +1,4 @@
-import '~/editor/styles.scss'
+import '~/toolbar/styles.scss'
 
 import React, { useRef } from 'react'
 import ReactDOM from 'react-dom'
@@ -7,8 +7,8 @@ import root from 'react-shadow'
 import Draggable from 'react-draggable'
 import { getContext, useActions, useValues } from 'kea'
 import { Provider } from 'react-redux'
-import { Toolbar } from '~/editor/Toolbar'
-import { dockLogic } from '~/editor/dockLogic'
+import { Toolbar } from '~/toolbar/Toolbar'
+import { dockLogic } from '~/toolbar/dockLogic'
 import { CloseOutlined } from '@ant-design/icons'
 import { initKea } from '~/initKea'
 import { useSecondRender } from 'lib/hooks/useSecondRender'
