+++ package.json
@@ -12,7 +12,7 @@
     },
     "homepage": "https://github.com/posthog/posthog#readme",
     "scripts": {
-        "copy-array": "cp node_modules/posthog-js/dist/array.js frontend/dist/",
+        "copy-array": "cp node_modules/@mariusandra/posthog-js/dist/array.js frontend/dist/",
         "start": "mkdir -p frontend/dist/ && cp -a frontend/public/* frontend/dist/ && npm run copy-array && webpack-dev-server --hotOnly",
         "start-https": "mkdir -p frontend/dist/ && cp -a frontend/public/* frontend/dist/ && npm run copy-array && webpack-dev-server --hotOnly --https",
         "build": "NODE_ENV=production webpack --config webpack.config.js && cp -a frontend/public/* frontend/dist/ && npm run copy-array",
@@ -37,7 +37,7 @@
         "kea-localstorage": "^1.0.2",
         "kea-router": "^0.3.0",
         "moment": "^2.24.0",
-        "posthog-js": "1.0.7-beta.0",
+        "@mariusandra/posthog-js": "1.1.0-beta.2",
         "prop-types": "^15.7.2",
         "react": ">= 16.8",
         "react-datepicker": "^2.13.0",
