+++ posthog/api/user.py
@@ -83,6 +83,7 @@ def redirect_to_site(request):
     }
     if settings.DEBUG:
         params['jsURL'] = 'http://localhost:8234/'
+        params['toolbarVersion'] = settings.TOOLBAR_VERSION
 
     state = urllib.parse.quote(json.dumps(params))
